Move sandbox related functions to sandbox.go
